requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
boto3==1.35.92
botocore==1.35.92
python-dotenv==1.0.1
